processors/workers/requirements.txt
stix2-validator==3.0.2
stix2==3.0.1
pymongo[srv]==4.2.0
python-dotenv==0.20.0